requirements.txt
Django==5.1.2
djangorestframework==3.15.2
python-dotenv==1.0.1
django-filter==24.3
psycopg2-binary==2.9.10
PyYAML==6.0.2
requests==2.32.3
djoser==2.2.3
celery==5.4.0
redis==5.2.0
django-baton==4.1.0
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
